Split the native-drop surfaces apart, and test the geometry
graph-native-drop.tsx was 1,327 lines: the sidebar tool insertion, the OS
file upload pipeline, a module-level drag signal, and two drop surfaces
whose hit-testing and indicator geometry were tangled into their DOM
measurement. It is gone; the three consumers import the pieces directly.

The point again is coverage. This app's vitest cannot parse .tsx, so none of
this had any — and that is not academic here: the flat-run boundary bug in
PL12-005, where the indicator pointed at one gap and the commit landed in
another, was a defect in exactly these functions, found by an e2e after it
shipped because nothing cheaper could reach it.

graph-native-drop-model.ts now holds the decisions with 41 tests, including
the grid indicator's four-case row-wrap geometry and stripDropAnchor itself.
mapWithConcurrency moves to lib/ with 7 more pinning its input-order and
concurrency-bound contracts.

resolveAnchoredTarget carried the same anchor-resolution rule TWICE, ten
lines apart and spelled differently: a nested ternary over
ids.indexOf(parseNodeId(...)) for the flat run, an if-ladder over
indexOfChildId for the children. Same three cases, same precedence, same
clamp — only the list differs, which is now the argument.

The two indicators are deliberately NOT shared. The strip anchors with
inset-y-1 and the grid with top-0, and both set `top`; one element carrying
both would be resolved by CSS source order. They share a class string and
state their own anchor, with the reason written down.

Behaviour is unchanged. Every comment travels with the code it explains.

Verified: tsc clean, 912 app tests (was 864), lint clean with the same 5
pre-existing <img> warnings, graph-view e2e 169/169.
